% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_templates.R
\name{create_template_interactive}
\alias{create_template_interactive}
\title{Create template interactively}
\usage{
create_template_interactive()
}
\value{
Character vector with template content
}
\description{
Create template interactively
}
\keyword{internal}
