% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_init.R
\name{get_user_template_dir}
\alias{get_user_template_dir}
\title{Get user template directory}
\usage{
get_user_template_dir()
}
\value{
Path to user template directory
}
\description{
Get user template directory
}
\keyword{internal}
