% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_setup.R
\name{is_banfa_setup}
\alias{is_banfa_setup}
\title{Check if banfa is set up}
\usage{
is_banfa_setup()
}
\value{
Logical. TRUE if at least one template exists.
}
\description{
Check if banfa is set up
}
\keyword{internal}
