% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_init.R
\name{load_template}
\alias{load_template}
\title{Load a project template}
\usage{
load_template(template_name)
}
\arguments{
\item{template_name}{Character string. Name of the template to load.}
}
\value{
Character vector containing the template content.
}
\description{
Load a project template
}
\keyword{internal}
