% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_setup.R
\name{rprj_setup}
\alias{rprj_setup}
\title{Set up banfa with initial templates}
\usage{
rprj_setup(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Should setup messages be displayed? Default is TRUE.}
}
\value{
Invisibly returns TRUE if setup was successful.
}
\description{
Creates the initial base template for banfa after installation.
This function should be run once after installing the package.
}
\examples{
\dontrun{
# Run after installing banfa
rprj_setup()
}
}
